requirements.txt
runpod
numpy
pillow
torch
torchvision
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git